requirements.txt
# SQL AI Agent
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
psycopg[binary]==3.2.3
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.6.1

# LLM (OpenAI-compatible API: Cursor, OpenAI, etc.)
openai==1.55.3

# Optional: for console with rich output
typer==0.15.1
rich==13.9.4
